refactor(scheduler): isolate ack decision boundary
